requirements.txt
amqp==2.5.2
appdirs==1.4.3
asgiref==3.2.3
attrs==19.3.0
billiard==3.6.2.0
black==19.10b0
celery==4.4.0
Click==7.0
coverage==5.0.3
dj-database-url==0.5.0
Django==3.0.4
django-celery-results==1.2.0
django-cors-headers==3.2.1
django-heroku==0.3.1
djangorestframework==3.11.0
factory-boy==2.12.0
Faker==4.0.1
gunicorn==20.0.4
importlib-metadata==1.5.0
isort==4.3.21
kombu==4.6.7
mypy==0.761
mypy-extensions==0.4.3
mysqlclient==1.4.6
pathspec==0.7.0
psycopg2==2.8.4
python-dateutil==2.8.1
pytz==2019.3
redis==3.4.1
regex==2020.2.20
six==1.14.0
sqlparse==0.3.0
text-unidecode==1.3
toml==0.10.0
typed-ast==1.4.1
typing-extensions==3.7.4.1
vine==1.3.0
whitenoise==5.0.1
zipp==3.0.0
